+++ classes/components/adapter/BaseAdapterInterface.class.php
@@ -1,0 +1,8 @@
+<?php
+namespace Flex\Components\Adapter;
+
+interface BaseAdapterInterface
+{
+    public function getVersion(): string;
+}
+?>